requirements.txt
asgiref==3.7.1
Django==4.2.1
django-material==1.12.0
django-material-widgets==1.0.0b3
django-tables2==2.5.3
djangorestframework==3.14.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3